backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.5.0 



